refactor: MemberService
