+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/MemberController.java
@@ -2,7 +2,6 @@ package com.startingblue.fourtooncookie.member;
 
 import com.startingblue.fourtooncookie.member.dto.request.MemberSaveRequest;
 import com.startingblue.fourtooncookie.member.dto.response.MemberSavedResponse;
-import com.startingblue.fourtooncookie.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
