+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/web/filter/AuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.startingblue.fourtooncookie.web.filter;
 
 import com.startingblue.fourtooncookie.web.exception.AuthenticationException;
-import com.startingblue.fourtooncookie.member.service.MemberService;
+import com.startingblue.fourtooncookie.member.MemberService;
 import com.startingblue.fourtooncookie.web.filter.jwt.JwtExtractor;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
