+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/global/authorization/member/MemberAdminAuthorizationInterceptorTest.java
@@ -1,6 +1,6 @@
 package com.startingblue.fourtooncookie.global.authorization.member;
 
-import com.startingblue.fourtooncookie.member.service.MemberService;
+import com.startingblue.fourtooncookie.member.MemberService;
 import com.startingblue.fourtooncookie.web.interceptor.MemberAdminAuthorizationInterceptor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
