+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/global/authorization/member/MemberSignedUpAuthorizationInterceptorTest.java
@@ -1,6 +1,6 @@
 package com.startingblue.fourtooncookie.global.authorization.member;
 
-import com.startingblue.fourtooncookie.member.service.MemberService;
+import com.startingblue.fourtooncookie.member.MemberService;
 import com.startingblue.fourtooncookie.web.interceptor.MemberSignedUpAuthorizationInterceptor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
